requirements.txt
asgiref==3.7.2
click==8.1.6
click-log==0.4.0
cloudinary==1.33.0
cloudinary-cli==1.7.1
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.20
django-allauth==0.54.0
django-cloudinary-storage==0.3.0
django-crispy-forms==1.14.0
django-phonenumber-field==7.1.0
docstring-parser==0.15
gunicorn==21.2.0
oauthlib==3.2.2
phonenumbers==8.13.18
psycopg2==2.9.6
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
requests-oauthlib==1.3.1
sqlparse==0.4.4
urllib3==1.26.15
whitenoise==6.5.0
